# src/TTAG.App/wwwroot/swagger/v1/swagger.json
{
...
  "swagger": "2.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "paths": {
...
    "/Users/Add": {
      "post": {
        "tags": [
          "Users"
        ],
        "operationId": "Users_Add",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/Users/Login": {
      "post": {
        "tags": [
          "Users"
        ],
        "operationId": "Users_Login",
        "parameters": [
          {
            "type": "string",
            "name": "username",
            "in": "query",
            "x-nullable": true
          },
          {
            "type": "string",
            "name": "password",
            "in": "query",
            "x-nullable": true
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    },
    "/Users": {
      "get": {
        "tags": [
          "Users"
        ],
        "operationId": "Users_GetUser",
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "query",
            "x-nullable": true
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    },
    "/Users/TestYourToken": {
      "get": {
        "tags": [
          "Users"
        ],
        "operationId": "Users_TestYourToken",
        "parameters": [
          {
            "type": "string",
            "name": "username",
            "in": "query",
            "x-nullable": true
          },
          {
            "type": "string",
            "name": "password",
            "in": "query",
            "x-nullable": true
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "Entity": {
      "type": "object",
      "required": [
        "createdDate",
        "lastUpdatedDate"
      ],
      "properties": {
        "id": {
          "type": "string"
        },
        "createdDate": {
          "type": "string",
          "format": "date-time"
        },
        "lastUpdatedDate": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
...
    "User": {
      "allOf": [
        {
          "$ref": "#/definitions/Entity"
        },
        {
          "type": "object",
          "properties": {
            "firstName": {
              "type": "string"
            },
            "lastName": {
              "type": "string"
            },
            "userName": {
              "type": "string"
            },
            "email": {
              "type": "string"
            },
            "password": {
              "type": "string"
            },
            "salt": {
              "type": "string",
              "format": "byte"
            }
          }
        }
      ]
    }
  }
}